refact(composables): extract fetchWithAuth from AgentActivityVisualization to useAgentActivityData (#6079)
